# A few additional minor corrections