--- a/src/Esprima/Ast/SourceType.cs
+++ b/src/Esprima/Ast/SourceType.cs
@@ -1,4 +1,4 @@
-﻿namespace Esprima;
+﻿namespace Esprima.Ast;
 
 public enum SourceType
 {