--- a/src/Esprima/Ast/SyntaxComment.cs
+++ b/src/Esprima/Ast/SyntaxComment.cs
@@ -3,7 +3,7 @@ using Esprima.Utils;
 
 namespace Esprima.Ast;
 
-public sealed class SyntaxComment : SyntaxElement
+public class SyntaxComment : SyntaxElement
 {
     public SyntaxComment(CommentType type, string value)
     {