--- a/src/Esprima/ParserException.cs
+++ b/src/Esprima/ParserException.cs
@@ -5,15 +5,19 @@ public sealed class ParserException : Exception
     public ParseError? Error { get; }
 
     public string? Description => Error?.Description;
+
     public string? SourceLocation => Error?.Source;
+
     /// <summary>
-    /// Zero-based index within the parsed code string. (Can be negative if code is not available.)
+    /// Zero-based index within the parsed code string. (Can be negative if location information is available.)
     /// </summary>
     public int Index => Error?.Index ?? -1;
+
     /// <summary>
-    /// One-based line number. (Can be zero if code is not available.)
+    /// One-based line number. (Can be zero if location information is not available.)
     /// </summary>
     public int LineNumber => Error?.LineNumber ?? 0;
+
     /// <summary>
     /// One-based column index.
     /// </summary>
@@ -23,19 +27,11 @@ public sealed class ParserException : Exception
     {
     }
 
-    public ParserException(string? message, Exception innerException) : this(message, null, innerException)
+    public ParserException(ParseError error) : this(null, null, error)
     {
     }
 
-    public ParserException(ParseError error) : this(null, error)
-    {
-    }
-
-    public ParserException(string? message, ParseError error) : this(message, error, null)
-    {
-    }
-
-    public ParserException(string? message, ParseError? error = null, Exception? innerException = null)
+    public ParserException(string? message, Exception? innerException = null, ParseError? error = null)
         : base(message ?? error?.ToString(), innerException)
     {
         Error = error;