--- a/src/Esprima/Utils/AstToJsonConverter.cs
+++ b/src/Esprima/Utils/AstToJsonConverter.cs
@@ -691,7 +691,7 @@ public class AstToJsonConverter : AstVisitor
 
             var callee = new ImportCompat
             {
-                Location = new Location(in import.Location.Start, new Position(import.Location.Start.Line, import.Location.Start.Column + importToken.Length)),
+                Location = new Location(import.Location.Start, new Position(import.Location.Start.Line, import.Location.Start.Column + importToken.Length)),
                 Range = new Range(import.Range.Start, import.Range.Start + importToken.Length)
             };
             var args = new NodeList<Expression>(new Expression[] { import.Source });