… test/Esprima.Tests/Fixtures/es2020/optional-chaining/optional-chaining-call-grouping.tree.json
… test/Esprima.Tests/Fixtures/es2020/optional-chaining/optional-chaining-computed-property-grouping.tree.json
… test/Esprima.Tests/Fixtures/es2020/optional-chaining/optional-chaining-new-grouping.tree.json
… test/Esprima.Tests/Fixtures/es2020/optional-chaining/optional-chaining-static-property-grouping.tree.json
… test/Esprima.Tests/LocationTests.cs
… test/Esprima.Tests/RegExpTests.cs
… test/Esprima.Tests/ScannerTests.cs