This PR started out as an enhancement of the sample project to make it (more) useful (#332). But along the way I found some missing parts to implement, issues to fix (and APIs to break... 😄) so I've decided that it's better off split into two parts.

BCs were made especially around error handling/reporting:
* IMO it doesn't make sense to collect exceptions in `CollectingErrorHandler` once we have a separate data class (`ParseError`) for describing errors. This would only result in unnecessary memory allocation since the exception class (`ParserException`) adds no value, just wraps the error class. I set this straight and added a method named `ToException` to convert error objects to throwable exception instances.
* There was a discrepancy between the definition of the type `Position` and its usage in `ParseError.Position`: it's documented that `Position` uses a zero-based column index. The type is used accordingly throughout the codebase except for `ParseError.Position`, which used one-based column indices. I fixed this but kept `ParseError.Column` one-based to not break everything unnecessary. I also documented this for clarity.

I also revised passing of `Position`, `Range` and `Location` parameters. Removed passing `Position` and `Range` structs by-ref as it doesn't really make sense to pass pointer-sized structs that way. On 64-bit platforms it just looks like an unnecessary indirection (unless inlined), plus it may increase the stack consumption of the caller method.

Of course, for the sake of certainty, I checked for regression in performance:

## Esprima.Benchmark.FileParsingBenchmark

| **Diff**|Method|FileName|Mean|Error|Gen 0|Gen 1|Gen 2|Allocated|
|------- |-------|-------|-------:|-------|-------:|-------:|-------:|-------:|
| Old |ParseProgram|angular-1.2.5|17.267 ms|0.0734 ms|750.0000|312.5000|-|4,315 KB|
| **New** |	|	| **17.106 ms (-1%)** | **0.0697 ms** | **750.0000 (0%)** | **312.5000 (0%)** | **-** | **4,315 KB (0%)** |
| Old |ParseProgram|backbone-1.1.0|2.322 ms|0.0103 ms|128.9063|35.1563|-|676 KB|
| **New** |	|	| **2.305 ms (-1%)** | **0.0160 ms** | **125.0000 (-3%)** | **46.8750 (+33%)** | **-** | **676 KB (0%)** |
| Old |ParseProgram|jquery-1.9.1|13.673 ms|0.0733 ms|640.6250|312.5000|-|3,764 KB|
| **New** |	|	| **13.587 ms (-1%)** | **0.0398 ms** | **625.0000 (-2%)** | **312.5000 (0%)** | **-** | **3,764 KB (0%)** |
| Old |ParseProgram|jquery.mobile-1.4.2|21.975 ms|0.1064 ms|968.7500|437.5000|-|5,798 KB|
| **New** |	|	| **21.659 ms (-1%)** | **0.2019 ms** | **1031.2500 (+6%)** | **500.0000 (+14%)** | **31.2500** | **5,798 KB (0%)** |
| Old |ParseProgram|mootools-1.4.5|10.938 ms|0.0175 ms|500.0000|250.0000|-|3,091 KB|
| **New** |	|	| **10.887 ms (0%)** | **0.0145 ms** | **500.0000 (0%)** | **250.0000 (0%)** | **-** | **3,091 KB (0%)** |
| Old |ParseProgram|underscore-1.5.2|1.946 ms|0.0118 ms|113.2813|46.8750|-|571 KB|
| **New** |	|	| **1.930 ms (-1%)** | **0.0093 ms** | **113.2813 (0%)** | **46.8750 (0%)** | **-** | **571 KB (0%)** |
| Old |ParseProgram|yui-3.12.0|9.830 ms|0.0397 ms|468.7500|234.3750|-|2,856 KB|
| **New** |	|	| **9.966 ms (+1%)** | **0.0367 ms** | **468.7500 (0%)** | **234.3750 (0%)** | **-** | **2,856 KB (0%)** |

With these changes we gained a few extra stack frames though: max. nesting depth has been increased from 585 to 589.