@adams85 2022-08-29
I was also strongly considering introducing a dedicated options class for `Scanner` because it's pretty weird that its public ctors accept a `ParserOptions` instance, even though that class contains options which are irrelevant to `Scanner`. We could do this without a severe BC if we created a `ScannerOptions` class or something and inherit `ParserOptions` from it. What do you think about this?