@adams85 2022-08-29
> Might as well to be separate things to have scanner and parser options (without inheritance). They are considered separate things during AST producing anyway.

Are you ok with [this solution] or you'd do it in another way (like composition instead of internal interface)?