@lahma 2022-08-29
I'm good, we can always refine if needed 😉